Models moved to shared module.
