+++ src/app/characters/pages/characters/character-list/character-card/character-card.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {Character} from "../../../../models/character";
+import {Character} from "../../../../../shared/models/character";
 
 @Component({
   selector: 'app-character-card',
