+++ src/app/shared/models/character.ts
