+++ src/app/shared/models/characterLocation.ts
