+++ src/app/shared/models/characterQuery.ts
